Install deps using Brew Bundle and provided Brewfile
The individual functions for installing packages were removed.  The
remaining were renamed.  Brew Bundle now installs these packages.  A new
function was also added to move the repo to the correct location so that
later symlinking doesn't require the user to clone the repo to the right
place.

TODO: Update the readme